# contracts/openapi/bff-api-v1.yaml
openapi: 3.1.1
info:
  title: NovaSteel BFF API
  version: 1.0.0
  description: |
    Versioned contract for the NovaSteel browser-facing BFF. All domain routes
    are mediated by the BFF; no browser route reaches Fabric, Foundry, ARM, or
    OT systems directly.
servers:
  - url: https://{host}
    variables:
      host:
        default: api.example.invalid
security:
  - EntraBearer: []
...
paths:
...
  /v1/devices:
    get:
      tags: [Devices]
      operationId: listDevices
      description: Device fleet projection, filtered to the caller's plant scope.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Query'
      responses:
        '200':
          description: Device fleet table.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /v1/devices/sensors/{sensorId}/series:
    get:
      tags: [Devices]
      operationId: getDeviceSensorSeries
      parameters:
        - $ref: '#/components/parameters/SensorId'
        - name: window
          in: query
          schema:
            type: string
            enum: ['15m', '1h', '4h', '8h', '24h']
            default: '1h'
        - name: points
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1440
            default: 120
      responses:
        '200':
          description: Down-sampled sensor time series for charting.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorSeriesEnvelope'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /v1/devices/incidents/{activeIncidentId}:
    delete:
      tags: [Devices]
      operationId: clearDeviceIncident
      parameters:
        - $ref: '#/components/parameters/ActiveIncidentId'
      responses:
        '200':
          description: Incident cleared; refreshed simulator status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulatorClearEnvelope'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /v1/devices/{deviceId}:
    get:
      tags: [Devices]
      operationId: getDevice
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Device detail with its full sensor list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDetailEnvelope'
        default:
          $ref: '#/components/responses/ErrorResponse'
components:
  securitySchemes:
    EntraBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Entra-issued user access token for the BFF audience.
  parameters:
...
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
    SensorId:
      name: sensorId
      in: path
      required: true
      schema:
        type: string
    ActiveIncidentId:
      name: activeIncidentId
      in: path
      required: true
      schema:
        type: string
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Sort:
      name: sort
      in: query
      schema:
        type: array
        items:
          type: string
          pattern: '^[A-Za-z][A-Za-z0-9_]*:(asc|desc)$'
      style: form
      explode: true
    Query:
      name: q
      in: query
      schema:
        type: string
        minLength: 1
  responses:
    ErrorResponse:
      description: Standard error envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    DataEnvelope:
      type: object
      additionalProperties: false
      required: [data, asOf, correlationId]
      properties:
        data:
          type: object
          additionalProperties: true
        asOf:
          type: string
          format: date-time
        correlationId:
          type: string
    ListEnvelope:
      type: object
      additionalProperties: false
      required: [items, total, page, size, asOf, correlationId]
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
          maximum: 200
        asOf:
          type: string
          format: date-time
        correlationId:
          type: string
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [code, message, correlationId, retryable]
      properties:
        code:
          type: string
          enum:
            - INVALID_TOKEN
            - FORBIDDEN_ROLE
            - FORBIDDEN_SCOPE
            - NOT_FOUND
            - VALIDATION_ERROR
            - IDEMPOTENCY_KEY_REQUIRED
            - IDEMPOTENCY_CONFLICT
            - STALE_APPROVAL
            - DUPLICATE_APPROVAL
            - RATE_LIMITED
            - UPSTREAM_UNAVAILABLE
            - CAPACITY_STATE_CONFLICT
            - ERASURE_STATE_CONFLICT
            - SIMULATOR_STATE_CONFLICT
            - POLICY_DENIED
            - INTERNAL_ERROR
        message:
          type: string
        correlationId:
          type: string
        retryable:
          type: boolean
...
    DeviceDetailEnvelope:
      allOf:
        - $ref: '#/components/schemas/DataEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/DeviceDetail'
    DeviceDetail:
      type: object
      additionalProperties: true
      required: [deviceId, site, area, description, status, sensorCount, healthScore, sensors]
      properties:
        deviceId:
          type: string
        site:
          type: string
        area:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [healthy, degraded, fault, offline]
        sensorCount:
          type: integer
          minimum: 0
        activeIncidents:
          type: array
          items:
            type: string
        lastSampleAt:
          type: [string, 'null']
          format: date-time
        healthScore:
          type: number
          minimum: 0
          maximum: 1
        uptimePct:
          type: number
        sensors:
          type: array
          items:
            $ref: '#/components/schemas/SensorRow'
    SensorRow:
      type: object
      additionalProperties: true
      required: [sensorId, deviceId, signalCode, displayName, unit, value, quality, status]
      properties:
        sensorId:
          type: string
        deviceId:
          type: string
        signalCode:
          type: string
        displayName:
          type: string
        area:
          type: string
        unit:
          type: string
        low:
          type: [number, 'null']
        high:
          type: [number, 'null']
        samplePeriodMs:
          type: integer
          description: 0 marks an event-driven signal rather than a periodic one.
        value:
          type: [number, 'null']
        quality:
          type: string
          enum: [good, uncertain, bad]
        status:
          type: string
          enum: [normal, warning, alarm, stale]
        trend:
          type: string
          enum: [rising, falling, flat]
        deviationPct:
          type: [number, 'null']
        clamped:
          type: boolean
        lastSampleAt:
          type: [string, 'null']
          format: date-time
    SensorSeriesEnvelope:
      allOf:
        - $ref: '#/components/schemas/DataEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SensorSeries'
    SensorSeries:
      type: object
      additionalProperties: true
      required: [sensorId, deviceId, displayName, unit, window, pointCount, points]
      properties:
        sensorId:
          type: string
        deviceId:
          type: string
        displayName:
          type: string
        unit:
          type: string
        low:
          type: [number, 'null']
        high:
          type: [number, 'null']
        window:
          type: string
        pointCount:
          type: integer
          minimum: 0
        points:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [t, v, q]
            properties:
              t:
                type: string
                format: date-time
              v:
                type: [number, 'null']
              q:
                type: string
                enum: [good, uncertain, bad]
...
    SimulatorStatus:
      type: object
      additionalProperties: true
      required:
        - state
        - scenario
        - seed
        - speedFactor
        - simulatedClock
        - tickCount
        - deviceCount
        - sensorCount
        - activeIncidents
      properties:
        state:
          type: string
          enum: [stopped, running, paused]
        scenario:
          type: string
        seed:
          type: integer
        speedFactor:
          type: number
        tickIntervalSeconds:
          type: number
        simulatedClock:
          type: string
          format: date-time
        elapsedHours:
          type: number
        tickCount:
          type: integer
        deviceCount:
          type: integer
        sensorCount:
          type: integer
        activeIncidents:
          type: array
          items:
            $ref: '#/components/schemas/ActiveIncident'
    ActiveIncident:
      type: object
      additionalProperties: true
      required: [activeIncidentId, incidentId, label, severity, deviceId, startedAt, endsAt]
      properties:
        activeIncidentId:
          type: string
        incidentId:
          type: string
        label:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        deviceId:
          type: string
        sensorId:
          type: [string, 'null']
        startedAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        remainingMinutes:
          type: number
        progress:
          type: number
          minimum: 0
          maximum: 1
...
    SimulatorClearEnvelope:
      allOf:
        - $ref: '#/components/schemas/DataEnvelope'
        - type: object
          properties:
            data:
              type: object
              additionalProperties: false
              required: [cleared, simulator]
              properties:
                cleared:
                  type: boolean
                simulator:
                  $ref: '#/components/schemas/SimulatorStatus'